+++ src/main/java/com/example/demo/service/impl/LoginEventServiceImpl.java
@@ -5,15 +5,15 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.example.demo1.entity.LoginEvent;
-import com.example.demo1.service.LoginEventService;
+import com.example.demo.entity.LoginEvent;
+import com.example.demo.service.LoginEventService;
 
 @Service
 public class LoginEventServiceImpl implements LoginEventService{
     List<LoginEvent> events =new ArrayList<>();
     long id=1;
      public LoginEvent recordLogin(LoginEvent event) {
-        event.setId((long) id++);
+        event.setId( id++);
         events.add(event);
         return event;
     }
